backend/requirements.txt
face-recognition
flask
flask-cors
pillow
